backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
google-api-python-client==2.70.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
openai==1.3.0
python-dotenv==1.0.0
psutil==5.9.4

# Optional - uncomment to enable local LLM support
# llama-cpp-python==0.2.11